Fix CI: clippy unnecessary cast, deny.toml config, Pages deploy
- Remove `as u32` cast on already-u32 value (clippy::unnecessary_cast)
- Rewrite deny.toml for cargo-deny v0.16+: remove invalid
  `unmaintained` field, remove "Ring" (not an SPDX license), add
  ring as a license exception, add GPL-2.0-or-later
- Add continue-on-error to Pages deploy (requires Pages to be
  enabled in repo Settings > Pages > Source: GitHub Actions)
- Docker build failure was the same clippy issue (exit code 101)

Copyright (C) 2026 Dyber, Inc.

+++ zernel-cli/src/commands/job.rs
@@ -298,7 +298,7 @@ pub async fn run(cmd: JobCommands) -> Result<()> {
                 .spawn()
                 .with_context(|| format!("failed to launch {launcher}. Is it installed?"))?;
 
-            let pid = child.id().unwrap_or(0) as u32;
+            let pid = child.id().unwrap_or(0);
 
             // Record in DB
             let conn = open_jobs_db()?;
